src/chat_with_pdf_func_impl/requirements.txt
azure-functions
azure-functions-durable
python-dotenv>=0.10.3
promptflow>=1.11.0
promptflow-tools

PyPDF2
faiss-cpu
openai
jinja2
tiktoken

## Telemetry
azure-monitor-opentelemetry
opentelemetry-instrumentation-openai>=0.21.3
# configure_azure_monitor installs the below with v0.42b0
# but we need >=0.45b0 for opentelemetry-semantic-conventions to enable openai
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-urllib>=0.45b0
opentelemetry-instrumentation-urllib3>=0.45b0
